+++ Makefile
@@ -3,7 +3,7 @@ githook:
 	ln -s -f ../../git_configs/hooks/pre-commit .git/hooks/pre-commit
 
 install: githook
-	pip install -e ocean pip-tools
+	pip install -e sdk/python pip-tools
 	pip-compile ocean/setup.py
 	pip-sync ocean/setup.txt requirements.txt
