+++ sdk/python/setup.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile
 # To update, run:
 #
-#    pip-compile --output-file ocean/setup.txt ocean/setup.py
+#    pip-compile --output-file sdk/python/setup.txt sdk/python/setup.py
 #
 appdirs==1.4.3
 aws-requests-auth==0.4.2
