+++ sdk/python/t4/api.py
@@ -374,7 +374,7 @@ def config(*autoconfig_url, **config_values):
     """Set or read the Helium configuration
 
     To retrieve the current config, call directly, without arguments:
-        >>> import helium as he
+        >>> import t4 as he
         >>> he.config()
 
     To trigger autoconfiguration, call with just the navigator URL:
@@ -384,7 +384,7 @@ def config(*autoconfig_url, **config_values):
         >>> he.config(navigator_url='http://example.com',
         ...           elastic_search_url='http://example.com/queries')
     When setting config values, unrecognized values are rejected.  Acceptable
-    config values can be found in `helium.util.CONFIG_TEMPLATE`
+    config values can be found in `t4.util.CONFIG_TEMPLATE`
 
     :param autoconfig_url: URL indicating a location to configure from
     :param **config_values: `key=value` pairs to set in the config
