+++ sdk/python/t4/snapshots.py
