+++ sdk/python/t4/util.py
