+++ sdk/python/tests/test_util.py
@@ -12,7 +12,7 @@ except ImportError: import pathlib
 import pytest
 
 ### Project imports
-from helium import util
+from t4 import util
 
 ### Constants
 TEST_YAML = """
