+++ sdk/python/tox.ini
@@ -5,7 +5,7 @@ requires =
 ; py27 -- it's pretty apparent 2.7 isn't a goal here, at least not yet.  Helium has a lot of py3 stuff.
 envlist = py35,py36
 
-; This is the "proper way" to install helium['tests'], but it doesn't work.
+; This is the "proper way" to install t4['tests'], but it doesn't work.
 extras = tests
 
 [venv]
